v-app-barを導入。スマホサイズへの適応はまだ
